+++ Example_MeliCardDrawer/Example_MeliCardDrawerTests/Tests/ControllerTests.swift
@@ -9,7 +9,7 @@
 import Foundation
 import XCTest
 @testable import Example_MeliCardDrawer
-@testable import MeliCardDrawer
+@testable import MLCardDrawer
 
 class CardHeaderControllerTest: XCTestCase {