+++ Source/Classes/Controller/CardHeaderController.swift
@@ -1,27 +1,40 @@
 import UIKit
 
 @objcMembers public class CardHeaderController: UIViewController {
+    private var shouldAnimate: Bool = true
     let cardFont = "RobotoMono-Regular"
     var frontView = FrontView()
     var backView = BackView()
 
+    required public init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
     public var cardUI: CardUI {
         willSet(value) {
-            frontView.setupAnimated(value)
+            if shouldAnimate {
+                frontView.setupAnimated(value)
+            } else {
+                frontView.setupUI(value)
+            }
             backView.setupUI(value)
         }
     }
 
-    public init(_ cardUI: CardUI, _ model: CardData) {
+    public init(_ cardUI: CardUI, _ model: CardData, _ disabledMode: Bool = false) {
         self.cardUI = cardUI
         UIFont.registerFont(fontName: cardFont, fontExtension: "ttf")
         super.init(nibName: nil, bundle: nil)
-        backView.setup(cardUI, model, view.frame)
-        frontView.setup(cardUI, model, view.frame)
+        backView.setup(cardUI, model, view.frame, disabledMode)
+        frontView.setup(cardUI, model, view.frame, disabledMode)
     }
 
     public func show() {
         transition(from: backView, to: frontView, .transitionFlipFromRight)
+    }
+
+    public func animated(_ animated: Bool) {
+        shouldAnimate = animated
     }
 
     public func showSecurityCode() {
@@ -40,7 +53,10 @@ import UIKit
         inView.addSubview(view)
         return self
     }
+}
 
+// MARK: Internals
+extension CardHeaderController {
     func transition(from origin: UIView, to destination: UIView, _ options: UIView.AnimationOptions) {
         addSubview(destination)
         Animator.flip(origin, destination, options)
@@ -49,9 +65,5 @@ import UIKit
     func addSubview(_ subview: UIView) {
         subview.frame = CGRect(origin: CGPoint.zero, size: view.frame.size)
         view.addSubview(subview)
-    }
-
-    required public init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
     }
 }