+++ Source/Classes/View/UIView+Nib.swift
@@ -1,7 +1,6 @@
 import UIKit
 
 internal extension UIView {
-
     func loadFromNib() {
         guard let view = Bundle(for: type(of: self)).loadNibNamed(String(describing: type(of: self)),
                                                   owner: self,