+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 v1.0
 🚀 First Release 🚀 
-- Stable 1.0 version
+- Stable 1.0 version iOS MLCardDrawer